+++ lib/LogSeverity.d.ts
@@ -3,6 +3,7 @@ export declare enum LogSeverity {
     DEBUG = 2,
     INFO = 4,
     WARNING = 8,
+    DEPRECATE = 8,
     ERROR = 16,
     FATAL = 32,
     ALL = 63