+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.20.0 (September 4, 2019)
+- Added `Logger.deprecate` method
+- Deprecate `StormError.getDetails`, use `StormError.getPrivateDetails` instead.
+- Deprecate `StormError.getAdditionalDetails`, use `StormError.getPublicDetails` instead.
+
+# 0.19.4 (August 27, 2019)
+- Updated dependency for vulnerability patches
+- Stricter ESLint rules
+- Updated UUID package
+
 # 0.19.3 (July 19, 2019)
 - Added "types" declaration in package.json